Updated documentation for common
